[add] [email] 파일 생성
